Refactoring, make buffer data location configurable.

+++ src/Sdk.Common/EtlDataPipeline/Nodes/Buffering/BufferRetrievalNode.cs
@@ -3,9 +3,18 @@ using Meshmakers.Octo.Sdk.Common.EtlDataPipeline.Nodes.Buffering.EdgeBuffer;
 
 namespace Meshmakers.Octo.Sdk.Common.EtlDataPipeline.Nodes.Buffering;
 
+/// <summary>
+/// Dummy configuration for the buffer retrieval node.
+/// This is only required for the infrastructure to work.
+/// Should never be set in a real pipeline configuration.
+/// </summary>
 [NodeName("BufferRetrievalNode", 1)]
 internal class BufferRetrievalNodeConfiguration : NodeConfiguration
 {
+    /// <summary>
+    /// Gets or sets a value indicating whether the data should be kept after sending.
+    /// </summary>
+    public bool? KeepDataAfterSending { get; set; }
 }
 
 [NodeConfiguration(typeof(BufferRetrievalNodeConfiguration))]
@@ -14,6 +23,9 @@ internal class BufferRetrievalNode(NodeDelegate next, IEdgeDataBuffer buffer) : 
     public async Task ProcessObjectAsync(IDataContext dataContext)
     {
         buffer.TryCloseCurrentChunk(true);
+
+        var config = dataContext.GetNodeConfiguration<BufferRetrievalNodeConfiguration>();
+        var shouldDelete = config.KeepDataAfterSending.GetValueOrDefault(false);
 
         foreach (var closedChunk in buffer.GetClosedChunks().ToList())
         {
@@ -26,7 +38,10 @@ internal class BufferRetrievalNode(NodeDelegate next, IEdgeDataBuffer buffer) : 
                 }
 
                 buffer.MarkAsSent(closedChunk);
-                buffer.DeleteChunk(closedChunk);
+                if (shouldDelete)
+                {
+                    buffer.DeleteChunk(closedChunk);
+                }
             }
             catch (Exception ex)
             {
